+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [1.0.8] - 2022-04-02
 ### Added
 - Add option to sort items when writing to file. [#29]
 - Add search/filter through cmd-f. [#64]